+++ lib/core_ext/hash.rb
