+++ spec/core_ext/hash_spec.rb
@@ -1,7 +1,6 @@
 # frozen_string_literal: true
 
 require 'spec_helper'
-require './lib/extensions/hash'
 
 describe 'hash extensions' do
   let(:keys) { %i[a b] }
